API

Modules

AffineMaps.AffineMaps

Types and constants

AffineMaps.AbstractAffineMap

AffineMaps.Add

AffineMaps.AddMul

AffineMaps.InvAddMul

AffineMaps.InvMul

AffineMaps.InvMulAdd

AffineMaps.Mul

AffineMaps.MulAdd

AffineMaps.Subtract

Functions and macros

Documentation

AffineMaps.AffineMaps — Module

AffineMaps

Implements affine transformations in a modular way.

AffineMaps.AbstractAffineMap — Type

abstract type AbstractAffineMap <: Function

Abstract type for affine maps.

Affine map f::AbstractAffineMap act like f(x) == A * x .+ b or f(x) == A * (x .+ b) or their inverses, depending on the type of the map f.

A may be a Number, AbstractMatrix{<:Number} or any other multiplicative linear operator in general (that supports at least size(A) and eltype(A)). The packages LinearMaps, LinearOperators and SciMLOperators provide such operators, for example.

b must have a shape that supports broadcasted addition and subtraction with x.

Subtypes of AbstractAffineMap should implement/support the APIs of

InverseFunctions.jl

ChangesOfVariables.jl

Functors.jl

AffineMaps.Add — Type

struct Add

f = Add(b) has the behavior f(x) == x .+ f.b.

See AbstractAffineMap for more information.

AffineMaps.AddMul — Type

struct AddMul

f = AddMul(A, b) has the behavior f(x) == f.A * (x .+ f.b).

See AbstractAffineMap for more information.

AffineMaps.InvAddMul — Type

muladd(f.A, x, f.b) struct InvAddMul

f = InvAddMul(A, b) has the behavior f(x) == (f.A \ x) .- f.b. It is the inverse of AddMul(A, b).

See AbstractAffineMap for more information.

AffineMaps.InvMul — Type

struct InvMul

f = InvMul(A) has the behavior f(x) == f.A \ x. It is the inverse of Mul(A).

See AbstractAffineMap for more information.

AffineMaps.InvMulAdd — Type

struct InvMulAdd

f = InvMulAdd(A, b) has the behavior f(x) == f.A \ (x .- f.b). It is the inverse of MulAdd(A, b).

See AbstractAffineMap for more information.

AffineMaps.Mul — Type

struct Mul

f = Mul(A) has the behavior f(x) == f.A * x.

See AbstractAffineMap for more information.

AffineMaps.MulAdd — Type

struct MulAdd

f = MulAdd(A, b) has the behavior f(x) == f.A * x .+ f.b.

See AbstractAffineMap for more information.

AffineMaps.Subtract — Type

struct Subtract

f = Subtract(b) has the behavior f(x) == x .- f.b. It is the inverse of Add(b).

See AbstractAffineMap for more information.